Split remaining >800-line source files into focused modules
Brings src/brain-renderer-webgl.js, src/routine-player.js, and
src/mini-routines.js under the 800-line file-size ceiling by extracting
self-contained pieces into new modules, re-exported/merged for identical
behavior:

- brain-renderer-webgl.js: GLSL sources moved to webgl-shaders.js,
  math/GL-setup helpers moved to webgl-gl-utils.js
- routine-player.js: CSV parsing moved to routine-csv.js, procedural
  routine generation moved to routine-procedural.js
- mini-routines.js: MINI_ROUTINES data split across
  mini-routines/part1.js and part2.js, merged via a barrel export
  (verified byte-for-byte equivalent to the original object)
