+++ src/brain-renderer-webgl.js
@@ -1,144 +1,9 @@
 import { BrainGeometry } from './brain-geometry.js';
 import { Mat4 } from './math-utils.js';
+import { meshVertexSource, meshFragmentSource, pointVertexSource, pointFragmentSource } from './webgl-shaders.js';
+import { clamp01, mix, smoothstep, createDefaultParams, createProgram } from './webgl-gl-utils.js';
 
 const BRAIN_RANGE = 1.6;
-
-const meshVertexSource = `#version 300 es
-precision highp float;
-layout(location = 0) in vec3 aPosition;
-layout(location = 1) in vec3 aColor;
-uniform mat4 uMvp;
-uniform float uPointSize;
-out vec3 vColor;
-void main() {
-    gl_Position = uMvp * vec4(aPosition, 1.0);
-    gl_PointSize = uPointSize;
-    vColor = aColor;
-}
-`;
-
-const meshFragmentSource = `#version 300 es
-precision highp float;
-in vec3 vColor;
-out vec4 outColor;
-void main() {
-    outColor = vec4(vColor, 1.0);
-}
-`;
-
-const pointVertexSource = `#version 300 es
-precision highp float;
-layout(location = 0) in vec3 aPosition;
-layout(location = 1) in vec4 aColorSize;
-uniform mat4 uMvp;
-out vec4 vColorSize;
-void main() {
-    gl_Position = uMvp * vec4(aPosition, 1.0);
-    gl_PointSize = max(1.5, aColorSize.w);
-    vColorSize = aColorSize;
-}
-`;
-
-const pointFragmentSource = `#version 300 es
-precision highp float;
-in vec4 vColorSize;
-out vec4 outColor;
-void main() {
-    vec2 uv = gl_PointCoord * 2.0 - 1.0;
-    float d = dot(uv, uv);
-    if (d > 1.0) {
-        discard;
-    }
-    float glow = smoothstep(1.0, 0.0, d);
-    outColor = vec4(vColorSize.rgb * glow, vColorSize.a * glow);
-}
-`;
-
-function clamp01(value) {
-    return Math.max(0, Math.min(1, value));
-}
-
-function mix(a, b, t) {
-    return a + (b - a) * t;
-}
-
-function smoothstep(edge0, edge1, x) {
-    const t = clamp01((x - edge0) / (edge1 - edge0 || 1e-6));
-    return t * t * (3 - 2 * t);
-}
-
-function createDefaultParams() {
-    return {
-        cognitiveLoad: 0.0,
-        frequency: 2.0,
-        amplitude: 0.5,
-        spikeThreshold: 0.8,
-        smoothing: 0.9,
-        style: 0.0,
-        sliceZ: 2.0,
-        flowSpeed: 4.0,
-        colorShift: 0.0,
-        sparkle: 0.0,
-        growth: 1.0,
-        shake: 0.0,
-        aberration: 0.0,
-        grain: 0.0,
-        focus: 0.5,
-        aperture: 0.0,
-        lightDirX: 1.0,
-        lightDirY: 1.0,
-        lightDirZ: 1.0,
-        ambientLight: 0.2,
-        dirIntensity: 0.8,
-        stress: 0.0,
-        cortisol: 0.0,
-        heavyMetal: 0.0,
-        myelin_degradation: 0.0,
-        fluidActive: 0.0,
-        immuneActivity: 0.0,
-        altitude: 0.0,
-        oxygenLevel: 1.0,
-        hypoxiaStress: 0.0,
-        metabolicRate: 1.0,
-        mitochondrialFunction: 1.0,
-        fogDensity: 0.0,
-        aiInfluence: 0.5,
-        resonanceThreshold: 0.2,
-        aiLayer: 0.0,
-        pointCloudDensity: 1.0,
-        fiberCoupling: 0.5,
-        foldScale: 1.0,
-        foldStrength: 0.16,
-        fissureDepth: 0.52,
-        lobeFoldBias: 1.0,
-        corticalThickness: 0.11,
-    };
-}
-
-function createProgram(gl, vertexSource, fragmentSource) {
-    const vertexShader = compileShader(gl, gl.VERTEX_SHADER, vertexSource);
-    const fragmentShader = compileShader(gl, gl.FRAGMENT_SHADER, fragmentSource);
-    const program = gl.createProgram();
-    gl.attachShader(program, vertexShader);
-    gl.attachShader(program, fragmentShader);
-    gl.linkProgram(program);
-    if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-        throw new Error(`WebGL program link failed: ${gl.getProgramInfoLog(program)}`);
-    }
-    gl.deleteShader(vertexShader);
-    gl.deleteShader(fragmentShader);
-    return program;
-}
-
-function compileShader(gl, type, source) {
-    const shader = gl.createShader(type);
-    gl.shaderSource(shader, source);
-    gl.compileShader(shader);
-    if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-        throw new Error(`WebGL shader compile failed: ${gl.getShaderInfoLog(shader)}`);
-    }
-    return shader;
-}
 
 export class BrainRendererWebGL {
     constructor(canvas) {
