+++ src/routine-player.js
@@ -10,6 +10,8 @@
 import { Easing, evaluateSpline } from './math-utils.js';
 import { CAMERA_PRESETS, handleCamera } from './routine-camera.js';
 import { createDefaultHandlers } from './routine-handlers.js';
+import { parseRoutineCSV } from './routine-csv.js';
+import { buildProceduralRoutine } from './routine-procedural.js';
 
 export class RoutinePlayer {
     // [Neuro-Script Cycle] WebGPU fallback logic confirmed
@@ -224,81 +226,7 @@ export class RoutinePlayer {
 
     // [Phase 8] Data Integration: CSV Parser
     parseCSV(text) {
-        const lines = text.split(/\r?\n/).filter(line => line.trim() !== '');
-        if (lines.length < 2) {
-            console.error("[Routine] CSV must have header and at least one row.");
-            return [];
-        }
-
-        const header = lines[0].split(',').map(h => h.trim().toLowerCase());
-        const rows = lines.slice(1);
-        const routine = [];
-
-        // Detect Format
-        if (header.includes('type')) {
-            console.log("[Routine] Detected CSV Event List format.");
-            // Format: time, type, key, value, duration, ease, message, target, intensity
-            rows.forEach(row => {
-                const cols = row.split(',').map(c => c.trim());
-                const event = {};
-
-                header.forEach((h, i) => {
-                    if (cols[i] !== undefined && cols[i] !== '') {
-                        // Type inference
-                        if (h === 'time' || h === 'duration' || h === 'intensity') {
-                            event[h] = parseFloat(cols[i]);
-                        } else if (h === 'value') {
-                             const f = parseFloat(cols[i]);
-                             event[h] = isNaN(f) ? cols[i] : f;
-                        } else {
-                            event[h] = cols[i];
-                        }
-                    }
-                });
-
-                if (event.time !== undefined && event.type) {
-                    routine.push(event);
-                }
-            });
-        } else {
-            console.log("[Routine] Detected CSV Time-Series format (fMRI).");
-            // Format: time, region1, region2...
-            // Check which headers map to regions
-            const regionIndices = {};
-            header.forEach((h, i) => {
-                // Check if header matches a known region
-                // We use this.regions which is passed in constructor
-                // or fall back to standard names
-                const knownRegions = this.regions ? Object.keys(this.regions) : [];
-                if (knownRegions.includes(h) || ['frontal', 'parietal', 'occipital', 'temporal', 'deep'].includes(h)) {
-                    regionIndices[i] = h;
-                }
-            });
-
-            rows.forEach(row => {
-                const cols = row.split(',').map(c => c.trim());
-                const timeIndex = header.indexOf('time');
-                if (timeIndex === -1) return;
-
-                const time = parseFloat(cols[timeIndex]);
-
-                if (!isNaN(time)) {
-                    Object.keys(regionIndices).forEach(idx => {
-                        const val = parseFloat(cols[idx]);
-                        if (!isNaN(val) && val > 0.05) { // Threshold
-                             routine.push({
-                                 time: time,
-                                 type: 'stimulus',
-                                 target: regionIndices[idx],
-                                 intensity: val * 5.0 // Scale for visibility (0.2 -> 1.0)
-                             });
-                        }
-                    });
-                }
-            });
-        }
-
-        return routine.sort((a, b) => a.time - b.time);
+        return parseRoutineCSV(text, this.regions);
     }
 
     loadRoutineFromCSV(text, loop = false) {
@@ -315,83 +243,7 @@ export class RoutinePlayer {
 
     // [Phase 3] Procedural Generation
     generateProceduralRoutine(duration = 30.0, intensity = 1.0) {
-        const routine = [];
-        const numEvents = Math.floor(duration * 0.5 * intensity); // ~1 event every 2 seconds
-
-        // Helper to pick random element
-        const pick = (arr) => arr[Math.floor(Math.random() * arr.length)];
-        const rand = (min, max) => Math.random() * (max - min) + min;
-
-        // Use available regions or defaults
-        const regions = Object.keys(this.regions).length > 0 ? Object.keys(this.regions) : ['frontal', 'occipital', 'parietal', 'temporal', 'deep'];
-        const styles = [0, 1, 2, 3];
-        const easeFuncs = Object.keys(Easing);
-
-        // Initial Setup
-        routine.push({ time: 0.0, type: 'text', message: 'Procedural Sequence Initiated', duration: 2.0 });
-        routine.push({ time: 0.0, type: 'style', value: pick(styles) });
-        routine.push({ time: 0.0, type: 'camera', target: 'global', duration: 2.0 });
-
-        let currentTime = 0.0;
-
-        for (let i = 0; i < numEvents; i++) {
-            currentTime += rand(1.0, 4.0);
-            if (currentTime > duration) break;
-
-            const eventType = pick(['stimulus', 'camera', 'style', 'lerp', 'text']);
-
-            if (eventType === 'stimulus') {
-                routine.push({
-                    time: currentTime,
-                    type: 'stimulus',
-                    target: pick(regions),
-                    intensity: rand(0.5, 1.5) * intensity
-                });
-            } else if (eventType === 'camera') {
-                 routine.push({
-                    time: currentTime,
-                    type: 'camera',
-                    target: pick(regions), // Use region names as camera targets
-                    duration: rand(1.0, 3.0),
-                    ease: pick(easeFuncs)
-                });
-            } else if (eventType === 'style') {
-                routine.push({
-                    time: currentTime,
-                    type: 'style',
-                    value: pick(styles)
-                });
-                routine.push({
-                     time: currentTime,
-                     type: 'text',
-                     message: 'Style Shift',
-                     duration: 1.0
-                });
-            } else if (eventType === 'lerp') {
-                const param = pick(['flowSpeed', 'amplitude', 'frequency', 'colorShift', 'sparkle']);
-                let val = 0;
-                if (param === 'flowSpeed') val = rand(1.0, 10.0);
-                if (param === 'amplitude') val = rand(0.2, 1.5);
-                if (param === 'frequency') val = rand(1.0, 8.0);
-                if (param === 'colorShift') val = rand(0.0, 1.0);
-                if (param === 'sparkle') val = rand(0.0, 1.0);
-
-                routine.push({
-                    time: currentTime,
-                    type: 'lerp',
-                    key: param,
-                    value: val,
-                    duration: rand(1.0, 3.0),
-                    ease: pick(easeFuncs)
-                });
-            }
-        }
-
-        // End with calm
-        routine.push({ time: duration, type: 'calm' });
-        routine.push({ time: duration, type: 'text', message: 'Sequence Complete', duration: 2.0 });
-
-        this.routine = routine.sort((a, b) => a.time - b.time);
+        this.routine = buildProceduralRoutine(this.regions, duration, intensity);
         this.loop = false; // Don't loop procedural routines by default
         this.stop(); // Reset player state
         console.log(`[Routine] Procedural Routine Generated (${this.routine.length} events)`);
